feat(config): add pydantic settings
