+++ server/cache.py
@@ -5,11 +5,11 @@ from typing import Any, Callable
 
 import redis
 
-from .config import Config
+from .settings import Settings
 
 __all__ = ["redis_cache", "clear_cache"]
 
-_redis = redis.Redis.from_url(Config().redis_url, decode_responses=True)
+_redis = redis.Redis.from_url(Settings().redis_url, decode_responses=True)
 
 
 def _make_key(prefix: str, args: tuple[Any, ...], kwargs: dict[str, Any]) -> str:
