+++ server/self_reflection.py
@@ -1,7 +1,7 @@
 from typing import Any
 
 
-from server.config import Config
+from server.settings import Settings
 from server.cache import redis_cache
 from loguru import logger
 from util import call_with_retries
@@ -13,7 +13,7 @@ __all__ = ["generate_self_critique"]
 def generate_self_critique(transcript: str, *, client: Any | None = None) -> str:
     """Return a short self-critique generated by an LLM."""
 
-    api_key = Config().openai_api_key
+    api_key = Settings().openai_api_key
     if client is None:
         if not api_key:
             return ""
@@ -28,7 +28,7 @@ def generate_self_critique(transcript: str, *, client: Any | None = None) -> str
     try:
         resp = call_with_retries(
             client.chat.completions.create,
-            model=Config().openai_model,
+            model=Settings().openai_model,
             messages=[
                 {
                     "role": "system",
