+++ tests/test_admin_api.py
@@ -33,7 +33,7 @@ import server.app as server_app  # noqa: E402
 server_app.generate_auth_url = _dummy_auth_url
 
 from server.app import create_app  # noqa: E402
-from server.config import Config  # noqa: E402
+from server.settings import Settings  # noqa: E402
 from fastapi.testclient import TestClient  # noqa: E402
 import server.state_manager as sm  # noqa: E402
 import fakeredis  # noqa: E402
@@ -51,7 +51,7 @@ def setup_app(monkeypatch, tmp_path):
     monkeypatch.setenv("OAUTH_CLIENT_ID", "cid")
     monkeypatch.setenv("OAUTH_AUTH_URL", "https://auth.example/authorize")
     key = db.create_api_key("tester")
-    app = create_app(Config())
+    app = create_app(Settings())
     client = TestClient(app)
     state = parse_qs(
         urlparse(
