+++ tests/test_async_calls.py
@@ -23,7 +23,7 @@ def test_async_inbound_call(monkeypatch, tmp_path):
 
     from server import app as server_app
     from server.app import create_app
-    from server.config import Config
+    from server.settings import Settings
 
     class DummyStateManager:
         def create_session(self, *a, **k):
@@ -46,7 +46,7 @@ def test_async_inbound_call(monkeypatch, tmp_path):
         server_app, "echo", types.SimpleNamespace(delay=lambda *_, **__: None)
     )
 
-    app = create_app(Config())
+    app = create_app(Settings())
     client = TestClient(app)
     resp = client.post(
         "/v1/inbound_call",
@@ -64,7 +64,7 @@ def test_async_recording_status(monkeypatch):
     monkeypatch.setenv("TOKEN_ENCRYPTION_KEY", base64.b64encode(b"0" * 16).decode())
     from server import app as server_app
     from server.app import create_app
-    from server.config import Config
+    from server.settings import Settings
 
     monkeypatch.setattr(server_app, "verify_api_key", lambda *_: True)
 
@@ -94,7 +94,7 @@ def test_async_recording_status(monkeypatch):
         types.SimpleNamespace(delay=fake_process),
     )
 
-    app = create_app(Config())
+    app = create_app(Settings())
     client = TestClient(app)
     resp = client.post(
         "/v1/recording_status",
