+++ tests/test_concurrency.py
@@ -12,7 +12,7 @@ import pytest
 import tests.test_api_key_auth  # noqa: F401
 from .db_utils import migrate_sqlite
 from server import app as server_app
-from server.config import Config
+from server.settings import Settings
 
 
 def test_concurrent_call_performance(
@@ -61,7 +61,7 @@ def test_concurrent_call_performance(
         server_app, "echo", types.SimpleNamespace(delay=lambda *_, **__: None)
     )
 
-    app = server_app.create_app(Config())
+    app = server_app.create_app(Settings())
 
     async def run_test() -> tuple[float, float]:
         transport = httpx.ASGITransport(app=app)
