+++ tests/test_config.py
@@ -16,7 +16,7 @@ os.environ.setdefault("TWILIO_AUTH_TOKEN", "token")
 os.environ.setdefault("TOKEN_ENCRYPTION_KEY", base64.b64encode(b"0" * 16).decode())
 
 
-from server.config import Config, ConfigError  # noqa: E402
+from server.settings import Settings, ConfigError  # noqa: E402
 from server.app import create_app  # noqa: E402
 
 
@@ -26,7 +26,7 @@ def test_config_missing_env(monkeypatch: pytest.MonkeyPatch) -> None:
     monkeypatch.delenv("TWILIO_ACCOUNT_SID", raising=False)
     monkeypatch.delenv("TWILIO_AUTH_TOKEN", raising=False)
     with pytest.raises(ConfigError):
-        Config()
+        Settings()
 
 
 def test_create_app_missing_env(monkeypatch: pytest.MonkeyPatch) -> None:
@@ -35,24 +35,24 @@ def test_create_app_missing_env(monkeypatch: pytest.MonkeyPatch) -> None:
     monkeypatch.delenv("TWILIO_ACCOUNT_SID", raising=False)
     monkeypatch.delenv("TWILIO_AUTH_TOKEN", raising=False)
     with pytest.raises(RuntimeError):
-        create_app(Config())
+        create_app(Settings())
 
 
 def test_create_app_missing_token_key(monkeypatch: pytest.MonkeyPatch) -> None:
     monkeypatch.delenv("TOKEN_ENCRYPTION_KEY", raising=False)
     with pytest.raises(RuntimeError):
-        create_app(Config())
+        create_app(Settings())
 
 
 def test_create_app_invalid_token_key(monkeypatch: pytest.MonkeyPatch) -> None:
     monkeypatch.setenv("TOKEN_ENCRYPTION_KEY", base64.b64encode(b"abc").decode())
     with pytest.raises(RuntimeError):
-        create_app(Config())
+        create_app(Settings())
 
 
 def test_config_embedding_model(monkeypatch: pytest.MonkeyPatch) -> None:
     monkeypatch.setenv("EMBEDDING_MODEL_NAME", "test-model")
     monkeypatch.setenv("EMBEDDING_PROVIDER", "openai")
-    cfg = Config()
+    cfg = Settings()
     assert cfg.embedding_model_name == "test-model"
     assert cfg.embedding_provider == "openai"
