+++ tests/test_language.py
@@ -14,7 +14,7 @@ install_vocode()
 
 
 from server import app as server_app  # noqa: E402
-from server.config import Config  # noqa: E402
+from server.settings import Settings  # noqa: E402
 from server import tasks  # noqa: E402
 from fastapi.testclient import TestClient  # noqa: E402
 from tools import language  # noqa: E402
@@ -64,7 +64,7 @@ def test_language_switch(monkeypatch: pytest.MonkeyPatch, tmp_path: Path) -> Non
 
     monkeypatch.setattr(server_app, "build_core_agent", fake_build_core_agent)
 
-    app = server_app.create_app(Config())
+    app = server_app.create_app(Settings())
     client = TestClient(app)
 
     from_num = "+15005550006"
