+++ tests/test_metrics.py
@@ -22,7 +22,7 @@ os.environ.setdefault("TOKEN_ENCRYPTION_KEY", base64.b64encode(b"0" * 16).decode
 
 
 from server.app import create_app  # noqa: E402
-from server.config import Config  # noqa: E402
+from server.settings import Settings  # noqa: E402
 from fastapi.testclient import TestClient  # noqa: E402
 
 
@@ -32,7 +32,7 @@ def test_metrics_endpoint(monkeypatch: pytest.MonkeyPatch, tmp_path: Path) -> No
     os.environ.setdefault("TWILIO_ACCOUNT_SID", "sid")
     os.environ.setdefault("TWILIO_AUTH_TOKEN", "token")
     os.environ.setdefault("TOKEN_ENCRYPTION_KEY", base64.b64encode(b"0" * 16).decode())
-    app = create_app(Config())
+    app = create_app(Settings())
     client = TestClient(app)
     db = migrate_sqlite(monkeypatch, tmp_path)
     key = db.create_api_key("tester")
