+++ tests/test_openapi_spec.py
@@ -12,7 +12,7 @@ from tests.utils.vocode_mocks import install as install_vocode
 install_vocode()
 
 from server.app import create_app  # noqa: E402
-from server.config import Config  # noqa: E402
+from server.settings import Settings  # noqa: E402
 
 
 def setup_client(monkeypatch, tmp_path):
@@ -24,7 +24,7 @@ def setup_client(monkeypatch, tmp_path):
     monkeypatch.setenv("TWILIO_AUTH_TOKEN", "token")
     monkeypatch.setenv("TOKEN_ENCRYPTION_KEY", base64.b64encode(b"0" * 16).decode())
     migrate_sqlite(monkeypatch, tmp_path)
-    app = create_app(Config())
+    app = create_app(Settings())
     transport = httpx.ASGITransport(app=app)
     client = httpx.AsyncClient(transport=transport, base_url="http://test")
     return client
